Calcualt the taxes with quantity and round up after. The code throws an exception if the quantity is negative.

+++ SalesTaxProblem/Domain/Models/ITransactionItem.cs
@@ -3,7 +3,7 @@
     public interface ITransactionItem
     {
         IProduct ProductPurchased { get; set; }
-        double Quantity { get; set; }
+        int Quantity { get; set; }
 
     }
 }
